backend/admin/requirements.txt
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
pyjwt>=2.8.0
boto3>=1.28.0